// HealthCare/src/Interfaces/CenterMaster/CenterResponse.ts
export interface CenterResponse {
  centerCode: string;
  centerName: string;
  centerInchargeName: string;
  salesIncharge: string;
  centerAddress: string;
  pincode: string;
  mobileNumber: string;
  alternateContactNo: string;
  emailId: string;
  rateType: string;
  centerStatus: boolean;
  introducedBy: string;
  creditLimit: number;
  isAutoBarcode: boolean;
  city: string;
  partnerId: string;
}